% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_ubesp_analysis}
\alias{create_ubesp_analysis}
\title{Create a UBESP analysis project}
\usage{
create_ubesp_analysis(
  path,
  affiliation = "ubesp",
  rstudio = rstudioapi::isAvailable(),
  open = interactive()
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist,
it is created, provided that the parent path exists.}

\item{affiliation}{the affiliaton for the project (default = "ubesp")}

\item{rstudio}{If \code{TRUE}, calls \code{\link[=use_rstudio]{use_rstudio()}} to make the new
supporting package and analysis project into \href{https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects}{RStudio Projects}.
If \code{FALSE} and a non-package project, a sentinel \code{.here} file is
placed so that the directory can be recognized as a project by the
\href{https://here.r-lib.org}{here} or
\href{https://rprojroot.r-lib.org}{rprojroot} packages.}

\item{open}{If \code{TRUE}, \link[=proj_activate]{activates} the new
analysis project:\preformatted{* If RStudio desktop, the package is opened in a new session.
* If on RStudio server, the current RStudio project is activated.
* Otherwise, the working directory and active project is changed.
}}
}
\value{
Path to the newly created project or package, invisibly.
}
\description{
This function create a standard UBESP analysis project:
* \code{create_ubesp_analisis()} creates a UBESP R analysis project

The function can be called on an existing project; you will be asked
before any existing files are changed.
}
\examples{
\dontrun{
    create_ubesp_analysis("~/test")
}
}
