% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devubesp-package.R
\docType{package}
\name{devubesp-package}
\alias{devubesp}
\alias{devubesp-package}
\title{devubesp: Automate Analyses Setup at UBESP}
\description{
Automate analyses setup tasks that are otherwise
    performed manually. This includes setting up directories, supporting
    packages and projects; it imports and expand the "usethis" package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CorradoLanera/devubesp}
  \item Report bugs at \url{https://github.com/CorradoLanera/devubesp/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} (0000-0002-0520-7428)

}
\keyword{internal}
