% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_data}
\alias{save_data}
\title{save_data}
\usage{
save_data(..., path, overwrite = FALSE, compress = "xz",
  compression_level = 9, version = 2)
}
\arguments{
\item{...}{the (unquoted) set of object to save}

\item{path}{the path in which to save the data}

\item{overwrite}{(lgl, default = FALSE) if TRUE and files with the
same names already exist, they were overwritten}

\item{compress}{the compression algoritm (default = xz)}

\item{compression_level}{the compresson level (default = 9)}

\item{version}{the workspace format version to use (see
\code{\link[base]{save}}).}
}
\description{
wrapper to \code{\link[base]{save}}).
}
